==> Scaden/Step_01_save_data_to anndata.ipynb <==
from tqdm import tqdm
import torch
import random
import numpy as np
import pandas as pd
import collections
import anndata
from anndata import read_h5ad
import scanpy as sc
from sklearn.preprocessing import MinMaxScaler

###       Load input data from a h5ad file and divide into training and test set 
#####     原始scaden好像沒有特別切分
#####     :param input_path: path to h5ad file
#####     :param batch_size: batch size to use for training
#####     :param datasets: a list of datasets to extract from the file
#####     :return: Dataset object

#input_path = "./pbmc_data.h5ad"
#raw_input = read_h5ad(input_path)

#raw_input
#raw_input.obs
#raw_input.var
#raw_input.X
#collections.Counter(raw_input.obs['ds'])

### Simulate RNA data from harmony 轉成 h5ad

exp = pd.read_csv("./0.input_data/0819_Simulate_RNA/0819_Simulate_RNA_after_harmony_filter.csv")
metadata = pd.read_csv("./0.input_data/0819_Simulate_RNA/0819-3.Simulate_RNA_after_harmony_filter_metadata.csv")
wall = pd.read_csv("./0.input_data/0819_Simulate_RNA/0819-1.Simulate_RNA_after_harmony_filter_wall_table.csv",
                   index_col=0)
exp = exp.drop(exp.columns[0],axis=1)

#gene = pd.read_csv("./0.input_data/0522_scRNA_25type_top1000_filter_genes.csv")
gene = pd.read_csv("./0.input_data/simulate_remain_cgl2_harmony_gene.csv")

print(exp.shape)
print(wall.shape)

symbol = list(gene['x'])
exp.index = symbol
#exp = exp.T

wall = wall.apply(lambda x: x/x.sum(), axis=1)

metadata.index = list(metadata['barcode'])
metadata = metadata['donor']
meta = pd.concat([wall, metadata], axis=1)

# log 2 transfromation
#genes_ID x sample
total_set = np.log2(exp+1)
total_set

# 對每個細胞進行 min max scaler
scaler = MinMaxScaler(feature_range=(0, 1))
total_set = scaler.fit_transform(total_set)
total_set = total_set.T
total_set.shape

adata = sc.AnnData(X = total_set, obs = meta, dtype = 'float32')
adata

adata.write('./0819_simulate_harmony_10p_400cell.h5ad')

print(adata.X)

collections.Counter(adata.obs['donor'])

adata.obs

### Simulate RNA data from SCT 轉成 h5ad

exp = pd.read_csv("./0.input_data/0819_Simulate_RNA/0819_Simulate_RNA_after_SCT_filter.csv")
metadata = pd.read_csv("./0.input_data/0819_Simulate_RNA/0819-3.Simulate_RNA_after_SCT_filter_metadata.csv")
wall = pd.read_csv("./0.input_data/0819_Simulate_RNA/0819-1.Simulate_RNA_after_SCT_filter_wall_table.csv",
                   index_col=0)
exp = exp.drop(exp.columns[0],axis=1)

#gene = pd.read_csv("./0.input_data/0622_scRNA_after_SCT_filter_genes_cv_20to200.csv")
gene = pd.read_csv("./0.input_data/simulate_remain_cgl2_sct_gene.csv")

print(exp.shape)
print(wall.shape)

symbol = list(gene['x'])
exp.index = symbol
#exp = exp.T

wall = wall.apply(lambda x: x/x.sum(), axis=1)

metadata.index = list(metadata['barcode'])
metadata = metadata['donor']
meta = pd.concat([wall, metadata], axis=1)

# log 2 transfromation
#genes_ID x sample
total_set = np.log2(exp+1)
total_set

# 對每個細胞進行 min max scaler
scaler = MinMaxScaler(feature_range=(0, 1))
total_set = scaler.fit_transform(total_set)
total_set = total_set.T
total_set.shape

adata = sc.AnnData(X = total_set, obs = meta, dtype = 'float32')
print(adata)

adata.write('./0819_simulate_SCT_cv20to200_10p_400cell.h5ad')

collections.Counter(adata.obs['donor'])

adata.obs
