==> Scaden/Step_03_save_cgl2_data_to_anndata.ipynb <==
## CGL2 in Harmony genes

from tqdm import tqdm
import torch
import random
import numpy as np
import pandas as pd
import collections
import anndata
from anndata import read_h5ad
import scanpy as sc
from sklearn.preprocessing import MinMaxScaler

exp = pd.read_csv("./0.input_data/CGL2_data/CGL2_harmony_matrix_1971.csv")
metadata = pd.read_csv("./0.input_data/CGL2_data/CGL2_metadata.csv")
#exp = exp.drop(exp.columns[0],axis=1)


gene = pd.read_csv("./0.input_data/CGL2_data/CGL2_harmony_genes_1971.csv")
print(gene.shape)

symbol = list(gene['x'])
exp.index = symbol

symbol = list(metadata['Scaden results'])
metadata.index = symbol
#exp = exp.T

# log 2 transfromation
#genes_ID x sample
total_set = np.log2(exp+1)
total_set

# 對每個細胞進行 min max scaler
scaler = MinMaxScaler(feature_range=(0, 1))
total_set = scaler.fit_transform(total_set)
total_set = total_set.T
total_set.shape

adata = sc.AnnData(X = total_set, obs = metadata, dtype = 'float32')
adata

adata.write('./0902_CGL2_harmony.h5ad')

## CGL2 in SCT genes

exp = pd.read_csv("./0.input_data/CGL2_data/CGL2_SCT_matrix_2870.csv")
metadata = pd.read_csv("./0.input_data/CGL2_data/CGL2_metadata.csv")
#exp = exp.drop(exp.columns[0],axis=1)

gene = pd.read_csv("./0.input_data/CGL2_data/CGL2_SCT_genes_2870.csv")
print(gene.shape)

symbol = list(gene['x'])
exp.index = symbol

symbol = list(metadata['Scaden results'])
metadata.index = symbol
#exp = exp.T

# log 2 transfromation
#genes_ID x sample
total_set = np.log2(exp+1)
total_set

# 對每個細胞進行 min max scaler
scaler = MinMaxScaler(feature_range=(0, 1))
total_set = scaler.fit_transform(total_set)
total_set = total_set.T
total_set.shape

adata = sc.AnnData(X = total_set, obs = metadata, dtype = 'float32')
adata

adata.write('./0902_CGL2_SCT.h5ad')
