==> Scaden/Step_04_predict_CG2.ipynb <==
## 在Github 找到有人建過 pytorch 版的 Scaden
###### https://github.com/poseidonchan/TAPE/blob/main/Experiments/pytorch_scaden_PBMConly.ipynb

!nvidia-smi

from tqdm import tqdm
import torch
import random
import numpy as np
import pandas as pd
import collections
import anndata
from anndata import read_h5ad
import torch.utils.data as Data
import torch.nn.functional as F
from sklearn.preprocessing import MinMaxScaler
from sklearn.model_selection import train_test_split
import torch.nn as nn
import matplotlib.pyplot as plt
import scanpy as sc
from torchsummary import summary
from scipy.stats import pearsonr
from sklearn import preprocessing

import os
os.environ["CUDA_VISIBLE_DEVICES"]="1"
device = torch.device("cuda" if torch.cuda.is_available() else "cpu")

adata = read_h5ad('./0902_CGL2_SCT.h5ad')
adata

adata = read_h5ad('./0902_CGL2_harmony.h5ad')
adata

#adata.obs

print(adata.X.shape)
adata.X[0,0:20]

## Scaden

class MLP_no_dropout(nn.Module):
    def __init__(self, gene_size, unit):
        super(MLP_no_dropout, self).__init__()
        self.gene_size = gene_size
        self.unit = unit
        
        self.D1 = nn.Sequential(
            nn.Linear(self.gene_size, 256),
            nn.ReLU(),
            nn.Linear(256, 128),
            nn.ReLU(),
            nn.Linear(128, 64),
            nn.ReLU(),
            nn.Linear(64, 32),
            nn.ReLU(),
            nn.Linear(32, self.unit),
            nn.Softmax(dim=1)
        )
        self.D2 = nn.Sequential(
            nn.Linear(self.gene_size, 512),
            nn.ReLU(),
            nn.Linear(512, 256),
            nn.ReLU(),
            nn.Linear(256, 128),
            nn.ReLU(),
            nn.Linear(128, 64),
            nn.ReLU(),
            nn.Linear(64, self.unit),
            nn.Softmax(dim=1)
        )
        self.D3 = nn.Sequential(
            nn.Linear(self.gene_size, 1024),
            nn.ReLU(),
            nn.Linear(1024, 512),
            nn.ReLU(),

            nn.Linear(512, 256),
            nn.ReLU(),
            nn.Linear(256, 128),
            nn.ReLU(),
            nn.Linear(128, self.unit),
            nn.Softmax(dim=1)
        )
    def forward(self, inputs):
        out1 = self.D1(inputs)
        out2 = self.D2(inputs)
        out3 = self.D3(inputs)
        
        out = (out1+out2+out3)/3
        return out

class MLP_batch(nn.Module):
    def __init__(self, gene_size, unit):
        super(MLP_batch, self).__init__()
        self.gene_size = gene_size
        self.unit = unit
        
        self.D1 = nn.Sequential(
            nn.Linear(self.gene_size, 256),
            nn.BatchNorm1d(256),
            nn.ReLU(),
            nn.Linear(256, 128),
            nn.BatchNorm1d(128),
            nn.ReLU(),
            nn.Linear(128, 64),
            nn.BatchNorm1d(64),
            nn.ReLU(),
            nn.Linear(64, 32),
            nn.BatchNorm1d(32),
            nn.ReLU(),
            nn.Linear(32, self.unit),
            nn.Softmax(dim=1)
        )
        self.D2 = nn.Sequential(
            nn.Linear(self.gene_size, 512),
            nn.BatchNorm1d(512),
            nn.ReLU(),
            nn.Linear(512, 256),
            nn.BatchNorm1d(256),
            nn.ReLU(),
            nn.Linear(256, 128),
            nn.BatchNorm1d(128),
            nn.ReLU(),
            nn.Linear(128, 64),
            nn.BatchNorm1d(64),
            nn.ReLU(),
            nn.Linear(64, self.unit),
            nn.Softmax(dim=1)
        )
        self.D3 = nn.Sequential(
            nn.Linear(self.gene_size, 1024),
            nn.BatchNorm1d(1024),
            nn.ReLU(),
            nn.Linear(1024, 512),
            nn.BatchNorm1d(512),
            nn.ReLU(),
            nn.Linear(512, 256),
            nn.BatchNorm1d(256),
            nn.ReLU(),
            nn.Linear(256, 128),
            nn.BatchNorm1d(128),
            nn.ReLU(),
            nn.Linear(128, self.unit),
            nn.Softmax(dim=1)
        )
    def forward(self, inputs):
        out1 = self.D1(inputs)
        out2 = self.D2(inputs)
        out3 = self.D3(inputs)
        
        out = (out1+out2+out3)/3
        return out

## Preprocess train, validation

### Training harmony

date = "0902"

dropout_rate = "no_dropout"

val_data = adata

#將 train_data 分開
#sample x gene
v_exp = val_data.X
#v_obs = val_data.obs
test = torch.tensor(v_exp, dtype=torch.float32)
#test_label = torch.tensor(np.array(v_obs), dtype=torch.float32)
celltype = ['CD141+DC','CD4/CD8-C1-CCR7','CD4/CD8-C2-MKI67','CD8-C7-KLRD1','CD8-C9-SLC4A10','Central memory T cells',
            'Circulating NK','Conventional dendritic cells(CD1C DC)','Cytotoxicity CD8T','DC-C4-LAMP3',
            'Effector memory T cells','Exhausted CD8+ T (Tex) cells','ILCs','Liver-resident NK (lrNK) cell',
            'Lymphoid-B','M-C4-GPX3','M1','Mast','Mono','Myeloid-derived suppressor cells','NK','TAM-like',
            'Th0','Th1','Treg']

PATH = './2.model/0902_model/0902_SCT_combine_model_no_dropout.pth'

mlp = MLP_no_dropout(v_exp.shape[1], len(celltype)).to(device)
mlp.load_state_dict(torch.load(PATH))

loss_fn = nn.L1Loss()
optimizer = torch.optim.Adam(mlp.parameters(),lr=0.0001)

test_dataset = Data.TensorDataset(test)
test_dataset = torch.utils.data.DataLoader(test_dataset, batch_size=64)

#print(mlp)

date = "0902"

dropout_rate = "batch_nor"

val_data = adata

#將 train_data 分開
#sample x gene
v_exp = val_data.X
#v_obs = val_data.obs
test = torch.tensor(v_exp, dtype=torch.float32)
#test_label = torch.tensor(np.array(v_obs), dtype=torch.float32)
celltype = ['CD141+DC','CD4/CD8-C1-CCR7','CD4/CD8-C2-MKI67','CD8-C7-KLRD1','CD8-C9-SLC4A10','Central memory T cells',
            'Circulating NK','Conventional dendritic cells(CD1C DC)','Cytotoxicity CD8T','DC-C4-LAMP3',
            'Effector memory T cells','Exhausted CD8+ T (Tex) cells','ILCs','Liver-resident NK (lrNK) cell',
            'Lymphoid-B','M-C4-GPX3','M1','Mast','Mono','Myeloid-derived suppressor cells','NK','TAM-like',
            'Th0','Th1','Treg']

PATH = './2.model/0902_model/0902_harmony_combine_model_batchnor_before.pth'

mlp = MLP_batch(v_exp.shape[1], len(celltype)).to(device)
mlp.load_state_dict(torch.load(PATH))

loss_fn = nn.L1Loss()
optimizer = torch.optim.Adam(mlp.parameters(),lr=0.0001)

test_dataset = Data.TensorDataset(test)
test_dataset = torch.utils.data.DataLoader(test_dataset, batch_size=64)

#print(mlp)

## Combine three submodel

train_loss_list = []
t_loss =[]
val_loss_list = []
v_loss = []

cor = []
cor_list = []
val_cor = []
val_cor_list = []

t_cor_list = []
v_cor_list = []
t_big_cor_list = []
v_big_cor_list = []

epochs = 100
big_type = [1,3,4,5,6,7,8,10,11,13,14,19,20,21,22,24]


#Train
for inte in range(epochs):
    print("--- Start training! ---", end="\r")
    print('      epoch: {}            '.format(inte+1))
    train_loss = 0
    val_loss = 0
    a = 0
    b = 0
    t_big = 0
    v_big = 0
    
#Validation    
    print("--- Now evaluation!!! ---",end="\r")
    mlp.eval()
    
    # Tell torch not to calculate gradients
    with torch.no_grad():
        for n, D in enumerate(test_dataset):
            pred = mlp(D[0].float().cuda())
            #VAL = loss_fn(pred, D[1].float().cuda())
            optimizer.zero_grad()
            optimizer.step()
            
        
print(" ----- Finish!! ----- ", end="\r")

model = "combine_model"

df = pd.DataFrame(pred.cpu().detach().numpy())
df = df*100
df

df.to_csv('./4.Result/0902_CGL2_SCT_reults.csv')

df.to_csv('./4.Result/0902_CGL2_harmony_reults.csv')
